% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_contribution.R
\name{acid_contribution}
\alias{acid_contribution}
\title{Calculate contributions of respiration and glycolysis to extracellular acid production}
\usage{
acid_contribution(PPR_total, OCR_total, OCR_nonmit, mho = 1, pH = 7.4,
  pK1 = 6.093)
}
\arguments{
\item{PPR_total}{Proton Production Rate = ECAR / buffering power, from Mookerjee et al. (2015b), equation 3.}

\item{OCR_total}{Oxygen Consumption Rate (pmol O2/min).}

\item{OCR_nonmit}{Non-mitochondrial OCR remaining after complete mitochondrial inhibition, eg with rotenone.}

\item{mho}{maximum H+/O2. 1.0 for glucose and glycogen, 0.8 for pyruvate, 0.65 for palmitate.}

\item{pH}{pH of the experiment.}

\item{pK1}{Overall pK for CO2(aq) + H2O -> HCO3− + H+, which is 6.093 at 37 Celsius.}
}
\description{
Calculate contributions of respiration and glycolysis to extracellular acid production from 
Mookerjee et al. (2015a), equation 3.
}
\references{
Mookerjee SA, Goncalves RL, Gerencser AA, Nicholls DG, Brand MD. The contributions of respiration and 
glycolysis to extracellular acid production. Biochim Biophys Acta. 2015 Feb;1847(2):171-81.

Mookerjee SA, Brand MD. Measurement and Analysis of Extracellular Acid Production to Determine 
Glycolytic Rate. J Vis Exp. 2015 Dec 12;(106):e53464.
}
\author{
Jonathan Dreyfuss
}
