% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/esstential_test.R
\name{essential_test}
\alias{essential_test}
\title{Block the specified reactions one at a time and maximiize the objective function}
\usage{
essential_test(model, reactions, solver = SYBIL_SETTINGS("SOLVER"))
}
\arguments{
\item{model}{A object of class \code{\link[sybil]{modelorg}}}

\item{reactions}{A character vector of reaction ids}

\item{solver}{The solver to use.  Default SYBIL_SETTINGS("SOLVER")}
}
\description{
This function iterates through each specified reaction, sets the bounds to lb=ub=0, and
maxmizes the objective function.
}
\author{
Alfred Ramirez
}

