% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxvar_rewrite.R
\name{fluxvar_rewrite}
\alias{fluxvar_rewrite}
\title{Rewrite fluxVar result into a matrix}
\usage{
fluxvar_rewrite(x)
}
\arguments{
\item{x}{A object of class \code{\link[sybil]{optsol}} produced by fluxVar}
}
\description{
This function rewrites an object of \code{\link[sybil]{optsol}} into a matrix
}
\references{
Ramirez AK, Lynes MD, Shamsi F, Xue R, Tseng YH, Kahn CR, Kasif S, Dreyfuss JM. Integrating Extracellular 
Flux Measurements and Genome-Scale Modeling Reveals Differences between Brown and White Adipocytes. Cell Rep 2017 
Dec; 21(11): 3040-3048.
}
\author{
Alfred Ramirez
}
