% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_reactions.R
\name{get_reactions}
\alias{get_reactions}
\title{Get all the reactions that contain a specified metabolite}
\usage{
get_reactions(metabolites, model)
}
\arguments{
\item{metabolites}{A character vector of metabolites}

\item{model}{A object of class \code{\link[sybil]{modelorg}}}
}
\description{
This function returns the stoichiometric matrix of all the reactions containing the specified metabolite(s)
}
\author{
Alfred Ramirez
}

