% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_seahorse.R
\name{map_seahorse}
\alias{map_seahorse}
\title{Map the seahorse fluxes to metabolic reactions}
\usage{
map_seahorse(x)
}
\arguments{
\item{x}{A matrix}
}
\description{
This function takes the matrix of sampled seahorse measurements returned by \code{\link{sample_seahorse}}
and maps the fluxes to metabolic reactions. It returns a matrix with the mapped fluxes of
basal oxygen consumption, mitochondrial oxygen consumption, mitochondrial ATP production,
mitochondrial ATP leak, and basal extracellular acidification
}
\author{
Alfred Ramirez, Jonathan Dreyfuss
}

