% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Recon_21A_reduced.R
\docType{data}
\name{Recon_21A_reduced}
\alias{Recon_21A_reduced}
\title{Recon 2.1A reduced model}
\format{An object of class \code{\link[sybil]{modelorg}}}
\usage{
data(Recon_21A_reduced)
}
\description{
Recon 2.1A with reactions removed whose bounds did not change in FVA with the integration of seahorse measurements
}
\keyword{datasets}

