% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_seahorse.R
\name{sample_seahorse}
\alias{sample_seahorse}
\title{Sample from the seahorse data}
\usage{
sample_seahorse(x, nsamples = 150, sample.nm = NULL)
}
\arguments{
\item{x}{A data.frame}

\item{nsamples}{The number of samples to generate}

\item{sample.nm}{The sample name}
}
\description{
This function takes the data.frame returned by the function \code{\link{summarize_seahorse}} and returns
A data.frame with sampled measurements for the specified sample
}
\author{
Alfred Ramirez
}

