% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert pmol/min/ug_protein or pmol/min/ug_DNA to mmol/gDw/hr, the units of the metabolic model}
\usage{
convert_units(x, mwr = 0.706)
}
\arguments{
\item{x}{A matrix}

\item{mwr}{Macromolecular weight ratio}
}
\description{
This function takes the matrix returned by \code{\link{sample_seahorse}} and multiplies it by a factor
}
\details{
The macromolecular weight ratio is the relationship between the initial normalization
(typically DNA or protein) and total dry weight. It can vary substantially depending on cell type. By default,
mwr = 0.706, which is the estimated protein/dry_weight ratio from the biomass reaction in Recon 2. For adipocytes,
mwr = 0.026
}
\author{
Alfred Ramirez
}

