% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxPredict.R
\name{fluxPredict}
\alias{fluxPredict}
\title{Make predictions from seahorse data by sampling and minimizing total flux}
\usage{
fluxPredict(model, seahorse_data, biomass_est = 0, alg = c("fba", "mtf"),
  obj_rxn = NULL, model.nm)
}
\arguments{
\item{model}{An object of class \code{\link[sybil]{modelorg}}}

\item{seahorse_data}{A data.frame returned by \code{\link{map_seahorse}}}

\item{biomass_est}{Estimated biomass flux. Default = 0.}

\item{alg}{Either "fba" (default) to optimize a reaction or "mtf" to only minimize total flux (and not do fba).}

\item{obj_rxn}{Reaction in \code{model} to optimize if \code{alg="fba"}. Default \code{NULL} optimizes ATP demand reaction.}

\item{model.nm}{The metabolic model name. One of "2.1A", "2.1x", or "2.2".}
}
\description{
This function integrates the sampled seahorse measurements as constraints into the specified model, 
optionally maximizes an objective reaction, minimizes total flux for each sample, and returns a matrix 
of reactions-by-samples where the entries are predicted fluxes.
}
\details{
For parallel computing, a parallel backend must be registered. See \code{\link[foreach]{foreach}} for details.
}
\author{
Alfred Ramirez, Jonathan Dreyfuss
}
