% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxvar_rewrite.R
\name{fluxvar_rewrite}
\alias{fluxvar_rewrite}
\title{Rewrite fluxVar result into a matrix}
\usage{
fluxvar_rewrite(x)
}
\arguments{
\item{x}{A object of class \code{\link[sybil]{optsol}} produced by fluxVar}
}
\description{
This function rewrites an object of \code{\link[sybil]{optsol}} into a matrix
}
\author{
Alfred Ramirez
}
