% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_seahorse.R
\name{sample_seahorse}
\alias{sample_seahorse}
\title{Sample from the seahorse data}
\usage{
sample_seahorse(x, sample.nm, nsamples = 150, nrep = 3)
}
\arguments{
\item{x}{A data.frame}

\item{sample.nm}{The sample name}

\item{nsamples}{The number of samples to generate}

\item{nrep}{The number of replicates, used to calculate standard error}
}
\description{
This function takes the data frame returned by the function \code{\link{summarize_seahorse}} and returns
a data frame with sampled measurements for the specified sample
}
\author{
Alfred Ramirez, Jonathan Dreyfuss
}
