% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_seahorse.R
\name{summarize_seahorse}
\alias{summarize_seahorse}
\title{Summarize seahorse data by sample and region}
\usage{
summarize_seahorse(x, n = 4)
}
\arguments{
\item{x}{A data.frame}

\item{n}{The number of regions}
}
\description{
This function takes the long format table exported by the XF Wave software for the 
oxidiative stress test and returns a matrix of means and standard deviations for each region
of the assay (namely basal, oligomycin, fccp, and rotenone).  The GroupName column must have the sample names.
The final measurement of each region is used for the mean and standard deviation across wells.
}
\author{
Alfred Ramirez
}
